+++ modules/avl-tree/test/test_sannikova_vika_avl_treee.cpp
@@ -1,90 +1,85 @@
 // Copyright 2018 Sannikova Vika
 
 #include <gtest/gtest.h>
-#include "include/avl-tree.h"
-#include <random>
-#include <chrono>
-#include <iostream>
-using namespace std;
-
+#include "include/avl_tree.h"
 
 TEST(Sannikova_Vika_AVLTreeTest, Can_Find_Min_In_AVL_Tree) {
-	// Arrange
-	AVL avltree;
-	int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
-	for (int i = 0; i < 10; i++) {
-		avltree.insert(avltree.getRoot(), numbers[i]);
-	}
+    // Arrange
+    AVL avltree;
+    int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
+    for (int i = 0; i < 10; i++) {
+        avltree.insert(avltree.getRoot(), numbers[i]);
+    }
 
-	// Act
-	int res = avltree.findMin(avltree.getRoot())->key;
+    // Act
+    int res = avltree.findMin(avltree.getRoot())->key;
 
-	// Assert
-	int exp_res = 0;
-	EXPECT_EQ(exp_res, res);
+    // Assert
+    int exp_res = 0;
+    EXPECT_EQ(exp_res, res);
 }
 
 TEST(Sannikova_Vika_AVLTreeTest, Can_Remove_Min_From_AVL_Tree) {
-	// Arrange
-	AVL avltree;
-	int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
-	for (int i = 0; i < 10; i++) {
-		avltree.insert(avltree.getRoot(), numbers[i]);
-	}
+    // Arrange
+    AVL avltree;
+    int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
+    for (int i = 0; i < 10; i++) {
+        avltree.insert(avltree.getRoot(), numbers[i]);
+    }
 
-	// Act
-	avltree.removeMin(avltree.getRoot());
+    // Act
+    avltree.removeMin(avltree.getRoot());
 
-	// Assert
-	int exp_res = 2;
-	EXPECT_EQ(exp_res, avltree.findMin(avltree.getRoot())->key);
+    // Assert
+    int exp_res = 2;
+    EXPECT_EQ(exp_res, avltree.findMin(avltree.getRoot())->key);
 }
 
 TEST(Sannikova_Vika_AVLTreeTest, Can_Find_Val_In_AVL_Tree) {
-	// Arrange
-	AVL avltree;
-	int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
-	for (int i = 0; i < 10; i++) {
-		avltree.insert(avltree.getRoot(), numbers[i]);
-	}
+    // Arrange
+    AVL avltree;
+    int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
+    for (int i = 0; i < 10; i++) {
+        avltree.insert(avltree.getRoot(), numbers[i]);
+    }
 
-	// Act
-	AVLNode* res = avltree.search(avltree.getRoot(), 91);
+    // Act
+    AVLNode* res = avltree.search(avltree.getRoot(), 91);
 
-	// Assert
-	AVLNode* exp_res = NULL;
-	EXPECT_NE(exp_res, res);
+    // Assert
+    AVLNode* exp_res = NULL;
+    EXPECT_NE(exp_res, res);
 }
 
 TEST(Sannikova_Vika_AVLTreeTest, Can_NOT_Find_Val_In_AVL_Tree) {
-	// Arrange
-	AVL avltree;
-	int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
-	for (int i = 0; i < 10; i++) {
-		avltree.insert(avltree.getRoot(), numbers[i]);
-	}
+    // Arrange
+    AVL avltree;
+    int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
+    for (int i = 0; i < 10; i++) {
+        avltree.insert(avltree.getRoot(), numbers[i]);
+    }
 
-	// Act
-	AVLNode* res = avltree.search(avltree.getRoot(), 55);
+    // Act
+    AVLNode* res = avltree.search(avltree.getRoot(), 55);
 
-	// Assert
-	AVLNode* exp_res = NULL;
-	EXPECT_EQ(exp_res, res);
+    // Assert
+    AVLNode* exp_res = NULL;
+    EXPECT_EQ(exp_res, res);
 }
 
 TEST(Sannikova_Vika_AVLTreeTest, Can_Remove_From_AVL_Tree) {
-	// Arrange
-	AVL avltree;
-	int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
-	for (int i = 0; i < 10; i++) {
-		avltree.insert(avltree.getRoot(), numbers[i]);
-	}
+    // Arrange
+    AVL avltree;
+    int numbers[10] = { 6, 4, 82, 0, 5, 42, 91, 2, 15, 90 };
+    for (int i = 0; i < 10; i++) {
+        avltree.insert(avltree.getRoot(), numbers[i]);
+    }
 
-	// Act
-	avltree.remove(avltree.getRoot(), 82);
-	AVLNode* res = avltree.search(avltree.getRoot(), 82);
+    // Act
+    avltree.remove(avltree.getRoot(), 82);
+    AVLNode* res = avltree.search(avltree.getRoot(), 82);
 
-	// Assert
-	AVLNode* exp_res = NULL;
-	EXPECT_EQ(exp_res, res);
+    // Assert
+    AVLNode* exp_res = NULL;
+    EXPECT_EQ(exp_res, res);
 }
